device refactor: move shared init logic into ForestDevice
